requirements.txt
fastapi[all]
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
langchain-google-genai
langchain
langchain-community
pandas
plotly
numpy